Refactor to RegisterReportService
